app/requirements.txt

mysql-connector